aml-investigation-ai/requirements.txt
# Core Dependencies
openai==1.54.3
pydantic==2.9.2
python-dotenv==1.0.1

# API Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Logging and Monitoring
loguru==0.7.2
rich==13.9.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Development
black==24.10.0
pylint==3.3.1
mypy==1.13.0
ipython==8.29.0
jupyter==1.1.1

# Optional: For production deployments
# gunicorn==23.0.0
# prometheus-client==0.21.0

